+++ src/transforms/shuffle.ts
@@ -5,6 +5,7 @@ import { computeProbabilityMap } from "../probability";
 import { getRandomInteger } from "../utils/random-utils";
 import Template from "../templates/template";
 import { Order } from "../order";
+import { isStaticValue } from "../utils/static-utils";
 
 export default ({ Plugin }: PluginArg): PluginObj => {
   const me = Plugin(Order.Shuffle);
@@ -17,7 +18,7 @@ export default ({ Plugin }: PluginArg): PluginObj => {
             return;
           }
           var illegalElement = path.node.elements.find((element) => {
-            !t.isStringLiteral(element);
+            !isStaticValue(element);
           });
 
           if (illegalElement) return;
@@ -28,12 +29,12 @@ export default ({ Plugin }: PluginArg): PluginObj => {
 
           var shift = getRandomInteger(
             1,
-            Math.min(40, path.node.elements.length * 6)
+            Math.min(30, path.node.elements.length * 6)
           );
 
           var shiftedElements = [...path.node.elements];
           for (var i = 0; i < shift; i++) {
-            shiftedElements.push(shiftedElements.shift());
+            shiftedElements.unshift(shiftedElements.pop());
           }
 
           var runtimeFn = me.getPlaceholder();