+++ test/transforms/shuffle.test.ts
@@ -1,13 +1,13 @@
 import JsConfuser from "../../src/index";
 
-it("should result in the same order", async () => {
+test("Variant #1: Result in the same order", async () => {
   var code = `
     var TEST_ARRAY = [1, 2, 3, 4, 5, 6, 7, 8, 9, 10];
 
     input(TEST_ARRAY);
   `;
 
-  var output = await JsConfuser(code, {
+  var { code: output } = await JsConfuser.obfuscate(code, {
     target: "browser",
     shuffle: true,
   });
@@ -21,12 +21,13 @@ it("should result in the same order", async () => {
 
   expect(value).toEqual([1, 2, 3, 4, 5, 6, 7, 8, 9, 10]);
 });
-it("should properly shuffle arrays within expressions", async () => {
+
+test("Variant #2: Properly shuffle arrays within expressions", async () => {
   var code = `
     input([1, 2, 3, 4, 5, 6, 7, 8, 9, 10]);
   `;
 
-  var output = await JsConfuser(code, {
+  var { code: output } = await JsConfuser.obfuscate(code, {
     target: "browser",
     shuffle: true,
   });
@@ -41,69 +42,18 @@ it("should properly shuffle arrays within expressions", async () => {
   expect(value).toEqual([1, 2, 3, 4, 5, 6, 7, 8, 9, 10]);
 });
 
-it("should shuffle arrays based on hash and unshuffle correctly", async () => {
-  var code = `input([1, 2, 3, 4, 5, 6, 7, 8, 9, 10]);`;
-
-  for (var i = 0; i < 20; i++) {
-    var output = await JsConfuser(code, {
-      target: "node",
-      shuffle: "hash",
-    });
-
-    var value;
-    function input(valueIn) {
-      value = valueIn;
-    }
-
-    eval(output);
-
-    expect(value).toEqual([1, 2, 3, 4, 5, 6, 7, 8, 9, 10]);
-  }
-});
-
-it("should shuffle arrays based on hash and unshuffle incorrect if changed", async () => {
-  var code = `input([1, 2, 3, 4, 5]);`;
-
-  var different = false;
-  for (var i = 0; i < 20; i++) {
-    var output = await JsConfuser(code, {
-      target: "node",
-      shuffle: "hash",
-    });
-
-    output =
-      output.split("[")[0] + "[" + output.split("[")[1].replace("5", "6");
-
-    var value;
-    function input(valueIn) {
-      value = valueIn;
-    }
-
-    eval(output);
-    expect(value).toHaveLength(5);
-    expect(typeof value[0] === "number").toStrictEqual(true);
-    if (value[0] !== 1) {
-      different = true;
-    }
-  }
-
-  expect(different).toStrictEqual(true);
-});
-
 // https://github.com/MichaelXF/js-confuser/issues/48
-it("Should properly apply to const variables", async () => {
+test("Variant #3: Properly apply to const variables", async () => {
   var code = `
       const TEST_ARRAY = [1, 2, 3, 4, 5, 6, 7, 8, 9, 10];
   
       input(TEST_ARRAY);
     `;
 
-  var output = await JsConfuser(code, {
+  var { code: output } = await JsConfuser.obfuscate(code, {
     target: "browser",
     shuffle: true,
   });
-
-  expect(output).toContain("TEST_ARRAY=function");
 
   var value;
   function input(valueIn) {
@@ -116,7 +66,7 @@ it("Should properly apply to const variables", async () => {
 });
 
 // https://github.com/MichaelXF/js-confuser/issues/53
-it("Should not use common variable names like x", async () => {
+test("Variant #4: Don't use common variable names like x", async () => {
   var code = `
       let x = -999;
       let a = [1, 2, 3, 4, 5, 6];
@@ -124,7 +74,7 @@ it("Should not use common variable names like x", async () => {
       VALUE = a;
     `;
 
-  var output = await JsConfuser(code, {
+  var { code: output } = await JsConfuser.obfuscate(code, {
     target: "browser",
     shuffle: true,
   });